+++ src/components/task.js
@@ -1,4 +1,5 @@
-import {MonthNames, formatTime} from '../utils/const';
+import {MonthNames} from '../utils/const';
+import {formatTime} from '../mock/task';
 
 const createHashtagsMarkup = (hashtags) => {
   return hashtags
@@ -15,7 +16,6 @@ const createHashtagsMarkup = (hashtags) => {
 };
 
 export const createTaskTemplate = (task) => {
-
   const {description, tags, dueDate, color, repeatingDays} = task;
 
   const isExpired = dueDate instanceof Date && dueDate < Date.now();
